\name{alumnadoEST}
\alias{alumnadoEST}
\docType{data}
\encoding{latin1}
\title{
Valoración del alumnado de la asignatura de Estadística
}
\description{
Datos acerca de la valoración de todo el material docente elaborado para la asignatura "Anàlisi de Dades en Psicologia"" por parte de una muestra de estudiantes del curso 2008-09. En concreto, se evaluó el grado de satisfacción respecto a diversos materiales docentes elaborados para esta asignatura. Se pretendía, además, relacionar estos resultados con otras variables, como son la calificación obtenida en la asignatura, el hecho de haber cursado previamente una asignatura niveladora (Fonaments Matemàtics). 
}
\usage{alumnadoEST}
\format{
  Una base de datos con 250 observaciones y  28 variables.
  \describe{
    \item{\code{identificador}}{Identificador de la persona encuestada}
    \item{\code{nota}}{Nota cualitativa de la persona}
    \item{\code{repetidor}}{Repite o no la asignatura}
    \item{\code{fundamentos.matematicos}}{Ha cursado o no la asignatura niveladora Fonaments Matemàtics}
    \item{\code{uso.libros.biblio}}{Ha usado o no las referencias propuestas en la bibliografía}
    \item{\code{uso.libros.prof}}{Ha usado o no las referencias elaboradas por el profesorado}
    \item{\code{uso.cuadro.resumen}}{Ha usado o no el cuadro resumen de la asignatura}
    \item{\code{uso.ppt}}{Ha usado o no las transparencias Power Point de la asignatura}
    \item{\code{uso.dosier}}{Ha usado o no el dosier de prácticas y sus soluciones para preparar la asignatura}
    \item{\code{uso.formulario}}{Ha usado o no el formulario}
    \item{\code{uso.glosario}}{Ha usado o no el glosario de la asignatura}
    \item{\code{uso.Pythia}}{Ha usado o no el recurso Pythia}
    \item{\code{uso.CDs}}{Ha usado o no los CDs interactivos de la asignatura}
    \item{\code{uso.clases.teor}}{Ha asistido o no a las sesiones teóricas}
    \item{\code{uso.tutorias}}{Ha asistido o no a las tutorías}
    \item{\code{uso.clases.prac}}{Ha asistido o no a las sesiones prácticas}
    \item{\code{libros.biblio}}{Grado de satisfacción respecto a las referencias propuestas en la bibliografía}
    \item{\code{libros.prof}}{Grado de satisfacción respecto a las referencias elaboradas por el profesorado}
    \item{\code{cuadro.resumen}}{Grado de satisfacción respecto al cuadro resumen de la asignatura}
    \item{\code{ppt}}{Grado de satisfacción respecto a las transparencias Power Point de la asignatura}
    \item{\code{dosier}}{Grado de satisfacción respecto al dosier de prácticas}
    \item{\code{formulario}}{Grado de satisfacción respecto al formulario}
    \item{\code{glosario}}{Grado de satisfacción respecto al glosario de la asignatura}
    \item{\code{Pythia}}{Grado de satisfacción respecto al recurso Pythia}
    \item{\code{CDs}}{Grado de satisfacción respecto a los CDs interactivos de la asignatura}
    \item{\code{teoria}}{Grado de satisfacción respecto a las sesiones teóricas}
    \item{\code{tutorias}}{Grado de satisfacción respecto a las tutorías}  
  }
}
\keyword{datasets}
