\name{fobia}
\alias{fobia}
\docType{data}
\encoding{latin1}
\title{
Eficacia tratamiento para la fobia
}
\description{
Datos simulados sobre la eficacia de un tratamiento psicológico en pacientes fóbicos.
}
\usage{fobia}
\format{
  Una base de datos con 160 observaciones y  14 variables.
  \describe{
    \item{\code{identif}}{Identificador numérico}
    \item{\code{sexo}}{Sexo del paciente}
    \item{\code{subtipo}}{Subtipo de fobia}
    \item{\code{edad}}{Edad del paciente}
    \item{\code{educa}}{Nivel educativo}
    \item{\code{detecci}}{Fase en la que se le detectó la enfermedad}
    \item{\code{grupo}}{Grupo al que se asigna al paciente}
    \item{\code{pensini}}{Pensamientos fóbicos al inicio de la investigación}
    \item{\code{pulsoini}}{Frecuencia cardíaca al inicio de la investigación}
    \item{\code{electini}}{Actividad electrodérmica al inicio de la investigación}
    \item{\code{ansiedad}}{Nivel de ansiedad del paciente}
    \item{\code{pensfin}}{Pensamientos fóbicos al finalizar la investigación}
    \item{\code{penseg}}{Medida de seguimiento respecto a los pensamientos fóbicos }
    \item{\code{pulsofin}}{Frecuencia cardíaca al finalizar la investigación}
    \item{\code{electfin}}{Actividad electrodérmica al finalizar la investigación}
  }
}

\keyword{datasets}
