\name{hoteles}
\alias{hoteles}
\docType{data}
\encoding{latin1}
\title{
Valoración del grado de satisfacción de los clientes de hoteles
}
\description{
Cuestionario acerca de la satisfacción de los clientes acerca del servicio ofrecido por los distintos departamentos de un hotel. La satisfacción se mide con ítems de tipo Likert en una escala graduada de 0 a 10 puntos.
}
\usage{hoteles}
\format{
  Una base de datos con 83 observaciones y  23 variables.
  \describe{
    \item{\code{identificador}}{Identificador de la persona encuestada}
    \item{\code{departamento}}{Departamento del hotel sobre el que se pasa la encuesta}
    \item{\code{idioma}}{Idioma de la persona encuestada}
    \item{\code{sexo}}{Sexo de la persona encuestada}
    \item{\code{edad}}{Edad en años cumplidos}
    \item{\code{nacionalidad}}{Nacionalidad de la persona encuestada}
    \item{\code{duracion.estancia}}{Duración de la estancia en el hotel}
    \item{\code{selec.hotel}}{Razón por la que se ha seleccionado el hotel}
    \item{\code{item1}}{Puntuación en el ítem 1 de la encuesta de satisfacción}
    \item{\code{item2}}{Puntuación en el ítem 2 de la encuesta de satisfacción}
    \item{\code{item3}}{Puntuación en el ítem 3 de la encuesta de satisfacción}
    \item{\code{item4}}{Puntuación en el ítem 4 de la encuesta de satisfacción}
    \item{\code{item5}}{Puntuación en el ítem 5 de la encuesta de satisfacción}
    \item{\code{item6}}{Puntuación en el ítem 6 de la encuesta de satisfacción}
    \item{\code{item7}}{Puntuación en el ítem 7 de la encuesta de satisfacción}
    \item{\code{item8}}{Puntuación en el ítem 8 de la encuesta de satisfacción}
    \item{\code{item9}}{Puntuación en el ítem 9 de la encuesta de satisfacción}
    \item{\code{item10}}{Puntuación en el ítem 10 de la encuesta de satisfacción}
    \item{\code{item11}}{Puntuación en el ítem 11 de la encuesta de satisfacción}
    \item{\code{item12}}{Puntuación en el ítem 12 de la encuesta de satisfacción}
    \item{\code{item13}}{Puntuación en el ítem 13 de la encuesta de satisfacción}
    \item{\code{item14}}{Puntuación en el ítem 14 de la encuesta de satisfacción}
    \item{\code{item15}}{Puntuación en el ítem 15 de la encuesta de satisfacción}
  }
}
\keyword{datasets}
