\name{memoriaVisual}
\alias{memoriaVisual}
\docType{data}
\encoding{latin1}
\title{
Memoria visual en una muestra de pacientes fóbicos
}
\description{
En un experimento se desea contrastar algunas características sobre la memoria visual y organización de los procesos al registrar y recuperar información en personas con algun trastorno fóbico. En una muestra de 30 sujetos con trastorno fóbico se ha registrado la variable número de imágenes recordadas, medida en dos momentos distintos: antes de un entrenamiento, el cual consiste en el adiestramiento en la organización de las imágenes, y después del entrenamiento. El objetivo es determinar si existen diferencias entre los dos momentos de medida. 
}
\usage{memoriaVisual}
\format{
  Una base de datos con 30 observaciones y  3 variables.
  \describe{
    \item{\code{antes}}{Número de imágenes recordadas antes de recibir el entrenamiento}
    \item{\code{despues}}{Número de imágenes recordadas después de recibir el entrenamiento}
    \item{\code{diferencia}}{Diferencias de los momentos antes y después en la muestra}  
  }
}
\keyword{datasets}
