\name{prevencionRL}
\alias{prevencionRL}
\docType{data}
\encoding{latin1}
\title{
Conducta Prevencionista Responsable
}
\description{
Estudio sobre el perfil de los trabajadores con conducta prevencionista responsable en el ámbito laboral.
}
\usage{prevencionRL}
\format{
  Una base de datos con 65 observaciones y  25 variables.
  \describe{
    \item{\code{grupo}}{Grupo al que se asignó a la persona}
    \item{\code{edad}}{Edad de la persona medida en años cumplidos}
    \item{\code{genero}}{Sexo de la persona encuestada}
    \item{\code{categoria.profesional}}{Categoría de la persona encuestada}
    \item{\code{nivel.estudios}}{Nivel de estudios de la persona encuestada}
    \item{\code{años.trabajados}}{Años trabajados por la persona}
    \item{\code{edad.inicio.trabajo}}{Edad en la que se empezó a trabajar}
    \item{\code{antiguedad.empresa}}{Antigüedad en la empresa en la que se trabaja actualmente}
    \item{\code{formacion.prl}}{Ha recibido o no formación en prevención de riesgos laborales}
    \item{\code{cps.est}}{Puntuación en el factor Estabilidad emocional del Cuestionario de Personalidad Situacional}
    \item{\code{cps.ans}}{Puntuación en el factor Ansiedad del Cuestionario de Personalidad Situacional}
    \item{\code{cps.auc}}{Puntuación en el factor Autoconcepto del Cuestionario de Personalidad Situacional}
    \item{\code{cps.efi}}{Puntuación en el factor Eficacia del Cuestionario de Personalidad Situacional}
    \item{\code{cps.csg}}{Puntuación en el factor Confianza/seguridad en si mismo del Cuestionario de Personalidad Situacional}
    \item{\code{cps.ind}}{Puntuación en el factor Independencia del Cuestionario de Personalidad Situacional}
    \item{\code{cps.dom}}{Puntuación en el factor Dominancia del Cuestionario de Personalidad Situacional}
    \item{\code{cps.ccg}}{Puntuación en el factor Control cognitivo del Cuestionario de Personalidad Situacional}
    \item{\code{cps.soc}}{Puntuación en el factor Sociabilidad del Cuestionario de Personalidad Situacional}
    \item{\code{cps.ajs}}{Puntuación en el factor Ajuste social del Cuestionario de Personalidad Situacional}
    \item{\code{cps.agr}}{Puntuación en el factor Agresividad del Cuestionario de Personalidad Situacional}
    \item{\code{cps.tol}}{Puntuación en el factor Tolerancia del Cuestionario de Personalidad Situacional}
    \item{\code{cps.ins}}{Puntuación en el factor Inteligencia social del Cuestionario de Personalidad Situacional}
    \item{\code{cps.inh}}{Puntuación en el factor Integridad/honestidad del Cuestionario de Personalidad Situacional}
    \item{\code{cps.lid}}{Puntuacion en el factor Liderazgo del Cuestionario de Personalidad Situacional}
    \item{\code{percentil.D48}}{Puntuación en percentil en el test de dominos D-48}
  }
}
\keyword{datasets}
