\name{RcmdrPlugin.EACSPIR-internal}
\title{RcmdrPlugin.EACSPIR: Una GUI para EACSPIR (funciones internas)}
\alias{ayuda.RcmdrPlugin.EACSPIR}
\alias{resumen.categoricas}
\alias{catResumen}
\alias{indices.categoricas}
\alias{catindices}
\alias{print.catindices}
\alias{grafico.Pareto}
\alias{pareto}
\alias{resumen.ordinales}
\alias{ordindices}
\alias{print.ordindices}
\alias{resumen.numericas}
\alias{numindices}
\alias{print.numindices}
\alias{histfun}
\alias{histograma}
\alias{bpfun}
\alias{diagrama.caja.ord}
\alias{diagrama.caja}
\alias{bivcat}
\alias{print.bivcat}
\alias{bivariante.categoricas}
\alias{barragfun}
\alias{barras.agrupadas}
\alias{mosaicofun}
\alias{grafico.mosaico}
\alias{bivord}
\alias{print.bivord}
\alias{bivariante.ordinales}
\alias{dispordfun}
\alias{dispersion.ordinales}
\alias{bivnum}
\alias{print.bivnum}
\alias{bivariante.numericas}
\alias{dispnumfun}
\alias{dispersion.numericas}
\alias{yuen.test}
\alias{print.yuen}
\alias{winvar}
\alias{yuenWelch.tTest}
\alias{UMWfun}
\alias{print.UMW}
\alias{U.Mann.Whitney}
\alias{TWfun}
\alias{print.TW}
\alias{pruebaT.Wilcoxon}
\alias{formafun}
\alias{print.Zforma}
\alias{prueba.conformidad.forma}
\alias{apuntfun}
\alias{print.Zapunt}
\alias{prueba.conformidad.apuntamiento}
\alias{tam.muestra}
\alias{print.tmuestra}
\alias{determ.tam.proporcion}
\alias{determ.tam.media}
\alias{normfun}
\alias{print.pnorm}
\alias{normgrfun}
\alias{pruebas.normalidad}
\alias{avarMRfun}
\alias{print.avarMR}
\alias{avarMRgraf}
\alias{avar.MR}
\encoding{latin1}
\description{Funciones internas (no deben llamarse mediante la consola de R) para el plugin de R-Commander de EACSPIR. 
  Para mas informacion consultar en \code{RcmdrPlugin.EACSPIR}.}
\keyword{internal}
