\name{RcmdrPlugin.EACSPIR-package}
\title{RcmdrPlugin.EACSPIR: Una GUI para un curso introductorio de Estadistica}
\alias{RcmdrPlugin.EACSPIR-package}
\alias{RcmdrPlugin.EACSPIR}
\docType{package}
\description{
  Este paquete proporciona una interfaz grafica de usuario (GUI) para poder obtener algunos de los indicadores, graficos y otras tecnicas estadisticas descritas en el libro "Estadistica Aplicada a las Ciencias Sociales mediante R y R-Commander". Esta GUI se instala dentro del paquete R-Commander.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.EACSPIR\cr
Type: \tab Package\cr
Version: \tab 0.2-2\cr
Date: \tab 2015-03-03\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Maribel Pero <mpero@ub.edu>
David Leiva <dleivaur@ub.edu>
Joan Guardia <jguardia@ub.edu>
Antonio Solanas <antonio.solanas@ub.edu>

Maintainer: David Leiva <dleivaur@ub.edu>
}

\keyword{package}
\seealso{
Se puede obtener mas informacion sobre R-Commander si se ejecuta \code{?Rcmdr} en la consola de R.
}

