\name{riesgosPS}
\alias{riesgosPS}
\docType{data}
\encoding{latin1}
\title{
Valoración de riesgos psicosociales en la empresa
}
\description{
Evaluación del riesgo psicosocial en una empresa espanola con el fin de conocer la salud laboral y psicosocial de la empresa. La evaluación se ha llevado a cabo a partir de la administración de la Batería MC-UB, que consta de un checklist a completar por el técnico de prevención de riesgos laborales de origen psicosocial, la realización de una entrevista a los directivos de la empresa y, finalmente, la administración de un cuestionario a los trabajadores de la empresa. En la presente base de datos se proporciona la información recogida a partir del cuestionario de la Batería MC-UB. Este cuestionario tiene en cuenta las siguientes áreas: organización del tiempo de trabajo, comunicación, formación y desarrollo, efectos sociales y de grupo, participación, contenido del trabajo y exigencia de la tarea y del entorno de trabajo.
}
\usage{riesgosPS}
\format{
  Una base de datos con 990 observaciones y  13 variables.
  \describe{
    \item{\code{identificador}}{Identificador de la persona evaluada}
    \item{\code{puesto}}{Categoría profesional de la persona}
    \item{\code{sexo}}{Sexo de la persona encuestada}
    \item{\code{edad}}{Edad en años cumplidos}
    \item{\code{{antiguedad.años}}}{Antigüedad en la empresa medida en años}
    \item{\code{antiguedad.meses}}{Antigüedad en la empresa medida en meses}
    \item{\code{organizacion.trabajo}}{Puntuación en el factor Organización del tiempo de trabajo de la batería MC-UB}
    \item{\code{comunicacion}}{Puntuación en el factor Comunicación de la batería MC-UB}
    \item{\code{formacion.desarrollo}}{Puntuación en el factor Formación y Desarrollo de la batería MC-UB}
    \item{\code{ef.sociales.grupo}}{Puntuación en el factor Efectos sociales y de grupo de la batería MC-UB}
    \item{\code{participacion}}{Puntuación en el factor Participación de la batería MC-UB}
    \item{\code{contenido.trabajo}}{Puntuación en el factor Contenido de trabajo de la batería MC-UB}
    \item{\code{efectos.tarea.entorno}}{Puntuación en el factor Exigencia de la tarea y del entorno de trabajo de la batería MC-UB}
  }
}
\keyword{datasets}
