\name{depresion}
\alias{depresion}
\docType{data}
\encoding{latin1}
\title{
Datos simulados sobre depresión
}
\description{
Datos simulados sobre depresión en tres poblaciones distintas.
}
\usage{depresion}
\format{
  Una base de datos con 21 observaciones y  2 variables.
  \describe{
    \item{\code{depresion}}{Nivel de depresión medido con una prueba psicométrica}
    \item{\code{poblacion}}{Población de referencia de los sujetos de la muestra}
  }
}

\keyword{datasets}
