\name{hoteles}
\alias{hoteles}
\docType{data}
\title{
Valoracion del grado de satisfaccion de los clientes de hoteles
}
\description{
Cuestionario acerca de la satisfaccion de los clientes acerca del servicio ofrecido por los distintos departamentos de un hotel. La satisfaccion se mide con items de tipo Likert en una escala graduada de 0 a 10 puntos.
}
\usage{hoteles}
\format{
  Una base de datos con 83 observaciones y  23 variables.
  \describe{
    \item{\code{identificador}}{Identificador de la persona encuestada}
    \item{\code{departamento}}{Departamento del hotel sobre el que se pasa la encuesta}
    \item{\code{idioma}}{Idioma de la persona encuestada}
    \item{\code{sexo}}{Sexo de la persona encuestada}
    \item{\code{edad}}{Edad en anos cumplidos}
    \item{\code{nacionalidad}}{Nacionalidad de la persona encuestada}
    \item{\code{duracion.estancia}}{Duracion de la estancia en el hotel}
    \item{\code{selec.hotel}}{Razon por la que se ha seleccionado el hotel}
    \item{\code{item1}}{Puntuacion en el item 1 de la encuesta de satisfaccion}
    \item{\code{item2}}{Puntuacion en el item 2 de la encuesta de satisfaccion}
    \item{\code{item3}}{Puntuacion en el item 3 de la encuesta de satisfaccion}
    \item{\code{item4}}{Puntuacion en el item 4 de la encuesta de satisfaccion}
    \item{\code{item5}}{Puntuacion en el item 5 de la encuesta de satisfaccion}
    \item{\code{item6}}{Puntuacion en el item 6 de la encuesta de satisfaccion}
    \item{\code{item7}}{Puntuacion en el item 7 de la encuesta de satisfaccion}
    \item{\code{item8}}{Puntuacion en el item 8 de la encuesta de satisfaccion}
    \item{\code{item9}}{Puntuacion en el item 9 de la encuesta de satisfaccion}
    \item{\code{item10}}{Puntuacion en el item 10 de la encuesta de satisfaccion}
    \item{\code{item11}}{Puntuacion en el item 11 de la encuesta de satisfaccion}
    \item{\code{item12}}{Puntuacion en el item 12 de la encuesta de satisfaccion}
    \item{\code{item13}}{Puntuacion en el item 13 de la encuesta de satisfaccion}
    \item{\code{item14}}{Puntuacion en el item 14 de la encuesta de satisfaccion}
    \item{\code{item15}}{Puntuacion en el item 15 de la encuesta de satisfaccion}
  }
}
\examples{
  data(hoteles)
  summary(hoteles)
}
\keyword{datasets}
