\name{prevencionRL}
\alias{prevencionRL}
\docType{data}
\title{
Conducta Prevencionista Responsable
}
\description{
Estudio sobre el perfil de los trabajadores con conducta prevencionista responsable en el ambito laboral.
}
\usage{prevencionRL}
\format{
  Una base de datos con 65 observaciones y  25 variables.
  \describe{
    \item{\code{grupo}}{Grupo al que se asigna a la persona}
    \item{\code{edad}}{Edad de la persona medida en anos cumplidos}
    \item{\code{genero}}{Sexo de la persona encuestada}
    \item{\code{categoria.profesional}}{Categoria de la persona encuestada}
    \item{\code{nivel.estudios}}{Nivel de estudios de la persona encuestada}
    \item{\code{anos.trabajados}}{Anos trabajados por la persona}
    \item{\code{edad.inicio.trabajo}}{Edad en la que se empezo a trabajar}
    \item{\code{antiguedad.empresa}}{Antiguedad en la empresa en la que se trabaja actualmente}
    \item{\code{formacion.prl}}{Ha recibido o no formacion en prevencion de riesgos laborales}
    \item{\code{cps.est}}{Puntuacion en el factor Estabilidad emocional del Cuestionario de Personalidad Situacional}
    \item{\code{cps.ans}}{Puntuacion en el factor Ansiedad del Cuestionario de Personalidad Situacional}
    \item{\code{cps.auc}}{Puntuacion en el factor Autoconcepto del Cuestionario de Personalidad Situacional}
    \item{\code{cps.efi}}{Puntuacion en el factor Eficacia del Cuestionario de Personalidad Situacional}
    \item{\code{cps.csg}}{Puntuacion en el factor Confianza/seguridad en si mismo del Cuestionario de Personalidad Situacional}
    \item{\code{cps.ind}}{Puntuacion en el factor Independencia del Cuestionario de Personalidad Situacional}
    \item{\code{cps.dom}}{Puntuacion en el factor Dominancia del Cuestionario de Personalidad Situacional}
    \item{\code{cps.ccg}}{Puntuacion en el factor Control cognitivo del Cuestionario de Personalidad Situacional}
    \item{\code{cps.soc}}{Puntuacion en el factor Sociabilidad del Cuestionario de Personalidad Situacional}
    \item{\code{cps.ajs}}{Puntuacion en el factor Ajuste social del Cuestionario de Personalidad Situacional}
    \item{\code{cps.agr}}{Puntuacion en el factor Agresividad del Cuestionario de Personalidad Situacional}
    \item{\code{cps.tol}}{Puntuacion en el factor Tolerancia del Cuestionario de Personalidad Situacional}
    \item{\code{cps.ins}}{Puntuacion en el factor Inteligencia social del Cuestionario de Personalidad Situacional}
    \item{\code{cps.inh}}{Puntuacion en el factor Integridad/honestidad del Cuestionario de Personalidad Situacional}
    \item{\code{cps.lid}}{Puntuacion en el factor Liderazgo del Cuestionario de Personalidad Situacional}
    \item{\code{percentil.D48}}{Puntuacion en percentil en el test de dominos D-48}
  }
}
\examples{
  data(prevencionRL)
  summary(prevencionRL)
}
\keyword{datasets}
