\name{RcmdrPlugin.EACSPIR-internal}
\title{RcmdrPlugin.EACSPIR: Una GUI para EACSPIR (funciones internas)}
\alias{ayuda.RcmdrPlugin.EACSPIR}
\alias{.chequeando.paquetes}
\alias{resumen.categoricas}
\alias{indices.categoricas}
\alias{grafico.Pareto}
\alias{resumen.ordinales}
\alias{resumen.numericas}
\alias{histograma}
\alias{diagrama.caja.ord}
\alias{diagrama.caja}
\alias{bivariante.categoricas}
\alias{barras.agrupadas}
\alias{grafico.mosaico}
\alias{bivariante.ordinales}
\alias{dispersion.ordinales}
\alias{bivariante.numericas}
\alias{dispersion.numericas}
\alias{yuen.test}
\alias{winvar}
\alias{yuenWelch.tTest}
\alias{U.Mann.Whitney}
\alias{pruebaT.Wilcoxon}
\alias{print.yuen}
\alias{prueba.conformidad.forma}
\alias{prueba.conformidad.apuntamiento}
\alias{tam.muestra}
\alias{print.tmuestra}
\alias{determ.tam.proporcion}
\alias{determ.tam.media}
\alias{pruebas.normalidad}
\alias{avar.MR}
\encoding{latin1}
\description{Funciones internas (no deben llamarse mediante la consola de R) para el plugin de R-Commander de EACSPIR. 
  Para mas información consultar en \code{RcmdrPlugin.EACSPIR}.}
\keyword{internal}
