\name{RcmdrPlugin.EACSPIR-package}
\title{RcmdrPlugin.EACSPIR: Una GUI para un curso introductorio de Estadística}
\alias{RcmdrPlugin.EACSPIR-package}
\alias{RcmdrPlugin.EACSPIR}
\docType{package}
\encoding{latin1}
\description{
  Este paquete proporciona una interfaz gráfica de usuario (GUI) para poder obtener algunos de los indicadores, gráficos y otras técnicas estadísticas descritos en el libro "Estadistica Aplicada a las Ciencias Sociales mediante el Programa R y R-Commander". Esta GUI se instala dentro del paquete R-Commander.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.EACSPIR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-03-07\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Maribel Peró <mpero@ub.edu>
David Leiva <dleivaur@ub.edu>
Joan Guàrdia <jguardia@ub.edu>
Antonio Solanas <antonio.solanas@ub.edu>

Maintainer: David Leiva <dleivaur@ub.edu>
}

\keyword{package}
\seealso{
Se puede obtener mas información sobre R-Commander si se ejecuta \code{?Rcmdr} en la consola de R.
}

