\name{testBarcelona}
\alias{testBarcelona}
\docType{data}
\encoding{latin1}
\title{
Datos del test Barcelona
}
\description{
Datos obtenidos mediante una versión reducida del test neuropsicológico Barcelona.
}
\usage{testBarcelona}
\format{
  Una base de datos con 499 observaciones y  25 variables.
  \describe{
    \item{\code{consulta}}{Tipo de servicio donde se realiza la evaluación}
    \item{\code{estado.civil}}{Estado civil de la persona evaluada}
    \item{\code{sexo}}{Sexo de la persona evaluada}
    \item{\code{derivado}}{Tipo de servicio que cursó la derivación}
    \item{\code{nivel.escolar}}{Nivel de estudios de la persona evaluada}
    \item{\code{ocupacion}}{Ocupación de la persona evaluada}
    \item{\code{dominancia}}{Descripción de la lateralidad de la persona evaluada}
    \item{\code{edad}}{Edad de la persona evaluada}
    \item{\code{anos.escolaridad}}{Años de escolaridad de la persona evaluada}
    \item{\code{lenguaje.espontaneo}}{Subtest del Test Barcelona que sirve para evaluar el lenguaje espontáneo de la persona}
    \item{\code{descripcion.lamina}}{Subtest del Test Barcelona que consiste en describir una lámina}
    \item{\code{fluencia}}{Subtest del Test Barcelona que sirve para evaluar la fluidez del lenguaje de la persona}
    \item{\code{contenido.informativo}}{Subtest del Test Barcelona}
    \item{\code{orientacion.persona}}{Evaluación de la orientación respecto a las personas}
    \item{\code{orientacion.lugar}}{Evaluación de la orientación respecto al lugar}
    \item{\code{orientacion.tiempo}}{Evaluación de la orientación respecto al tiempo}
    \item{\code{digitos.directos}}{Subtest del Test Barcelona que consiste en recordar en el mismo orden una secuencia de dígitos}
    \item{\code{digitos.inversos}}{Subtest del Test Barcelona que consiste en recordar en orden inverso una secuencia de dígitos}
    \item{\code{series.directas}}{Subtest del Test Barcelona que consiste en recordar en el mismo orden una secuencia de palabras}
    \item{\code{series.directas.tiempo}}{Subtest del Test Barcelona que consiste en recordar en el mismo orden una secuencia de palabras. La prueba tiene un límite de tiempo}
    \item{\code{series.inversas}}{Subtest del Test Barcelona que consiste en recordar en orden inverso una secuencia de palabras}
    \item{\code{series.inversas.tiempo}}{Subtest del Test Barcelona que consiste en recordar en orden inverso una secuencia de palabras. La prueba tiene un límite de tiempo}
    \item{\code{repeticion.logatomos}}{Subtest del Test Barcelona}
    \item{\code{repeticion.palabras}}{Subtest del Test Barcelona que consiste en repetir palabras}
    \item{\code{diagnostico}}{Diagnóstico médico realizado}
  }
}
\keyword{datasets}
